refactor: order패키지 구조 변경
1. 컨벤션에 맞춰 패키지 생성하여 Order, OrderStatus 위치 변경
2. FranchiseStore에 apiKey 컬럼 생성
