+++ src/main/java/com/fourweekdays/fourweekdays/order/model/entity/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.fourweekdays.fourweekdays.order;
+package com.fourweekdays.fourweekdays.order.model.entity;
 
 public enum OrderStatus {
     REQUESTED, APPROVED, SHIPPED, CANCELLED
